+++ package.json
@@ -83,7 +83,7 @@
     "isotope-packery": "2.x.x",
     "jquery": "^2.2.4",
     "lodash": "4.15.0",
-    "openlayers": "4.3.1",
+    "ol": "^4.4.2",
     "perfect-scrollbar": "0.7.x",
     "proj4": "2.3.3",
     "promise-queue": "2.2.3",
