+++ web/index.html
@@ -104,37 +104,7 @@ Licensed under the NASA Open Source Agreement, Version 1.3 (http://opensource.gs
                       <link type="text/css" rel="Stylesheet" href="pages/css/wv.document.css">
 
 
-    <!-- External libraries =============================================== -->
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/babel-polyfill/dist/polyfill.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/babel-polyfill/dist/polyfill.min.js"></script> -->
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/react/dist/react.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/react/dist/react.min.js"></script> -->
-
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/react-dom/dist/react-dom.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/react-dom/dist/react-dom.min.js"></script> -->
-
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/worldview-components/browser/wvc.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/worldview-components/browser/wvc.min.js?v=@BUILD_NONCE@"></script> -->
-
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/lodash/lodash.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/lodash/lodash.min.js"></script> -->
-
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/bluebird/js/browser/bluebird.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/bluebird/js/browser/bluebird.min.js"></script> -->
-
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/promise-queue/lib/index.js"></script>
-
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/openlayers/dist/ol-debug.js"></script>
-    <!-- link.dev -->       <script type="text/javascript" src="js/ol.wv/ol.wv.lookupimagetile.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="js/ol.js"></script> -->
-
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/gifshot/build/custom/gifshot.custom.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/gifshot/build/custom/gifshot.custom.min.js"></script> -->
-
-    <!-- link.dev -->       <script type="text/javascript" src="ext/node_modules/clipboard/dist/clipboard.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/clipboard/dist/clipboard.min.js"></script> -->
-
-
+  
     <!-- link.dev  -->      <script type="text/javascript" src="ext/main/jquery-2.1.4/jquery.js"></script>
     <!-- link.prod --> <!-- <script type="text/javascript" src="ext/main/jquery-2.1.4/jquery.min.js"></script> -->
                             <script type="text/javascript" src="ext/main/jquery.migrate-1.2.1/jquery-migrate.min.js"></script>
@@ -150,9 +120,6 @@ Licensed under the NASA Open Source Agreement, Version 1.3 (http://opensource.gs
 
     <!-- link.dev  -->      <script type="text/javascript" src="ext/ui/jquery.jscrollpane-2.0.22/jquery.jscrollpane.js"></script>
     <!-- link.prod --> <!-- <script type="text/javascript" src="ext/ui/jquery.jscrollpane-2.0.22/jquery.jscrollpane.min.js"></script> -->
-
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/perfect-scrollbar/dist/js/perfect-scrollbar.jquery.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/perfect-scrollbar/dist/js/perfect-scrollbar.jquery.min.js"></script> -->
 
     <!-- link.dev  -->      <script type="text/javascript" src="ext/date/mobiscroll-2.6.0/mobiscroll.js"></script>
     <!-- link.prod --> <!-- <script type="text/javascript" src="ext/date/mobiscroll-2.6.0/mobiscroll.min.js"></script> -->
@@ -171,16 +138,10 @@ Licensed under the NASA Open Source Agreement, Version 1.3 (http://opensource.gs
 
     <!-- link.dev  -->      <script type="text/javascript" src="ext/ui/d3/d3.js"></script>
     <!-- link.prod --> <!-- <script type="text/javascript" src="ext/ui/d3/d3.min.js"></script> -->
-
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/isotope-layout/dist/isotope.pkgd.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/isotope-layout/dist/isotope.pkgd.min.js"></script> -->
-    <!-- link.dev  -->      <script type="text/javascript" src="ext/node_modules/isotope-packery/packery-mode.pkgd.js"></script>
-    <!-- link.prod --> <!-- <script type="text/javascript" src="ext/node_modules/isotope-packery/packery-mode.pkgd.min.js"></script> -->
-
     <!-- link.dev  -->      <script type="text/javascript" src="ext/layers/jcarousel/jcarousel.js"></script>
     <!-- link.prod --> <!-- <script type="text/javascript" src="ext/layers/jcarousel/jcarousel.min.js"></script> -->
 
-    <!-- link.dev --> <!-- <script type="text/javascript" src="js/wv.js?v=@BUILD_NONCE@&mobile"></script> -->
+    <!-- link.dev --> <script type="text/javascript" src="build/wv.js"></script>
     <!-- link.prod --> <!-- <script type="text/javascript" src="js/wv.min.js?v=@BUILD_NONCE@&mobile"></script> -->
 
 </head>
